Extract caches and concepts.
